fix b64 encoding/decoding using utils from shotgun; prevents buffer overflow
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@76005 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ PROTO/azy/src/include/Azy.h
@@ -311,7 +311,7 @@ EAPI void                      azy_rpc_log_enable(void);
 
 /* utils */
 EAPI char                     *azy_base64_encode(const char *string, double len);
-EAPI char                     *azy_base64_decode(const char *string, int len);
+EAPI char                     *azy_base64_decode(const char *string, size_t len);
 EAPI unsigned char            *azy_memstr(const unsigned char *big, const unsigned char *small, size_t big_len, size_t small_len);
 EAPI const char               *azy_uuid_new(void);
 EAPI Azy_Net_Transport         azy_transport_get(const char *content_type);
