+++ PROTO/azy/src/lib/extras/cencode.c
@@ -23,7 +23,7 @@ char base64_encode_value(char value_in)
         return encoding[(int)value_in];
 }
 
-int base64_encode_block(const char* plaintext_in, int length_in, char* code_out, base64_encodestate* state_in)
+int base64_encode_block(const char* plaintext_in, size_t length_in, char* code_out, base64_encodestate* state_in)
 {
         const char* plainchar = plaintext_in;
         const char* const plaintextend = plaintext_in + length_in;
@@ -71,13 +71,14 @@ int base64_encode_block(const char* plaintext_in, int length_in, char* code_out,
                         *codechar++ = base64_encode_value(result);
                         result  = (fragment & 0x03f) >> 0;
                         *codechar++ = base64_encode_value(result);
-
+/*
                         ++(state_in->stepcount);
                         if (state_in->stepcount == CHARS_PER_LINE/4)
                         {
                                 *codechar++ = '\n';
                                 state_in->stepcount = 0;
                         }
+*/
                 }
         }
         /* control should not reach here */
